refactor(formatters): remove unused Granite32/33 processor classes (#1304)
The Granite32InputProcessor, Granite32OutputProcessor, Granite32ChatCompletion,
Granite33InputProcessor, Granite33OutputProcessor, and Granite33ChatCompletion
classes were carried over from granite-common (#571) but never wired into any
backend or session routing. Mellea operates at the action/context level via
intrinsic processors; the chat-message-level Granite3x pipeline was never used.

Deletes the entire mellea/formatters/granite/granite3/ subtree (15 files),
removes the six re-exported symbols from the top-level granite/__init__.py,
and deletes the five test files that covered only this dead code. The
nltk_check tests (base.optional) from the shared test file are rescued into
test/formatters/granite/base/test_base_optional.py before deletion.

Domain owner sign-off: frreiss confirmed removal in issue #1304.
Closes #1304.
