--- mellea/formatters/granite/granite3/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-"""Input and output processing for the Granite 3 family of models."""
-
-# Local
-from .types import Granite3Controls
-
-__all__ = ("Granite3Controls",)
