--- mellea/formatters/granite/granite3/granite32/__init__.py
@@ -1,14 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-"""Input and output processing for the Granite 3.2 family of models."""
-
-# Local
-from .input import Granite32InputProcessor
-from .output import Granite32OutputProcessor
-from .types import Granite32ChatCompletion
-
-__all__ = (
-    "Granite32ChatCompletion",
-    "Granite32InputProcessor",
-    "Granite32OutputProcessor",
-)
