--- mellea/formatters/granite/granite3/granite33/__init__.py
@@ -1,14 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-"""Input and output processing for the Granite 3.3 family of models."""
-
-# Local
-from .input import Granite33InputProcessor
-from .output import Granite33OutputProcessor
-from .types import Granite33ChatCompletion
-
-__all__ = (
-    "Granite33ChatCompletion",
-    "Granite33InputProcessor",
-    "Granite33OutputProcessor",
-)
